refactor: move quote totals to widget
