+++ lib/screens/simplified_quote_screen.dart
@@ -16,6 +16,7 @@ import '../widgets/quote_type_selector.dart';
 import '../widgets/main_product_selection.dart';
 import '../widgets/permits_section.dart';
 import '../widgets/tax_rate_section.dart';
+import '../widgets/quote_totals_section.dart';
 import '../controllers/quote_form_controller.dart';
 
 class SimplifiedQuoteScreen extends StatefulWidget {
@@ -629,7 +630,15 @@ class _SimplifiedQuoteScreenState extends State<SimplifiedQuoteScreen> {
                 _autoDetectTaxRate(context.read<AppStateProvider>()),
           ),
           const SizedBox(height: 16),
-          _buildQuoteTotalsSection(),
+          QuoteTotalsSection(
+            quoteLevels: _quoteLevels,
+            mainProduct: _mainProduct,
+            mainQuantity: _mainQuantity,
+            taxRate: _taxRate,
+            permits: _permits,
+            customLineItems: _customLineItems,
+            quoteType: _quoteType,
+          ),
         ],
       ],
     );
@@ -638,568 +647,6 @@ class _SimplifiedQuoteScreenState extends State<SimplifiedQuoteScreen> {
 
 
 
-  Widget _buildQuoteTotalsSection() {
-    final currencyFormat = NumberFormat.currency(symbol: '\$');
-
-    return Card(
-      elevation: 3,
-      color: _quoteType == 'multi-level' ? Colors.blue.shade50 : Colors.green.shade50,
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Icon(
-                  Icons.calculate,
-                  color: _quoteType == 'multi-level' ? Colors.blue.shade700 : Colors.green.shade700,
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  _quoteType == 'multi-level' ? 'Quote Totals' : 'Quote Total',
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    fontWeight: FontWeight.bold,
-                    color: _quoteType == 'multi-level' ? Colors.blue.shade800 : Colors.green.shade800,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 16),
-
-            if (_quoteType == 'multi-level') ...[
-              // Multi-level layout (existing)
-              Column(
-                children: _quoteLevels.map((level) {
-                  // Calculate totals including permits and custom items
-                  final levelSubtotal = level.subtotal;
-                  final permitsTotal = _permits.fold(0.0, (sum, permit) => sum + permit.amount);
-                  final taxableCustomItems = _customLineItems.where((item) => item.isTaxable).fold(0.0, (sum, item) => sum + item.amount);
-                  final nonTaxableCustomItems = _customLineItems.where((item) => !item.isTaxable).fold(0.0, (sum, item) => sum + item.amount);
-
-                  final taxableSubtotal = levelSubtotal + permitsTotal + taxableCustomItems;
-                  final nonTaxableSubtotal = nonTaxableCustomItems;
-                  final totalSubtotal = taxableSubtotal + nonTaxableSubtotal;
-                  final taxAmount = taxableSubtotal * (_taxRate / 100);
-                  final totalWithTax = totalSubtotal + taxAmount;
-
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 12),
-                    padding: const EdgeInsets.all(12),
-                    decoration: BoxDecoration(
-                      color: Colors.white,
-                      borderRadius: BorderRadius.circular(8),
-                      border: Border.all(color: Colors.blue.shade300),
-                    ),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          level.name,
-                          style: TextStyle(
-                            fontWeight: FontWeight.bold,
-                            fontSize: 16,
-                            color: Colors.blue.shade800,
-                          ),
-                        ),
-                        const SizedBox(height: 8),
-
-                        // Main product
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            Expanded(
-                              child: Text(
-                                '${_mainProduct!.name} (${_mainQuantity.toStringAsFixed(1)} ${_mainProduct!.unit})',
-                                style: const TextStyle(fontWeight: FontWeight.w500),
-                              ),
-                            ),
-                            Text(
-                              currencyFormat.format(level.baseProductTotal),
-                              style: const TextStyle(fontWeight: FontWeight.bold),
-                            ),
-                          ],
-                        ),
-
-                        // Added products
-                        // Added products
-                        ..._addedProducts.map((product) => Padding(
-                          padding: const EdgeInsets.only(top: 4),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Expanded(
-                                child: Text(
-                                  '${product.productName} (${product.quantity.toStringAsFixed(1)} ${product.unit})',
-                                ),
-                              ),
-                              Text(currencyFormat.format(product.totalPrice)),
-                            ],
-                          ),
-                        )),
-
-// NEW: Show permits if any
-                        if (_permits.isNotEmpty) ...[
-                          const SizedBox(height: 8),
-                          Container(
-                            padding: const EdgeInsets.all(6),
-                            decoration: BoxDecoration(
-                              color: Colors.orange.shade50,
-                              borderRadius: BorderRadius.circular(4),
-                            ),
-                            child: Column(
-                              children: [
-                                Row(
-                                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    Text(
-                                      'PERMITS:',
-                                      style: TextStyle(
-                                        fontWeight: FontWeight.w600,
-                                        color: Colors.orange.shade800,
-                                        fontSize: 11,
-                                      ),
-                                    ),
-                                    Text(
-                                      currencyFormat.format(permitsTotal),
-                                      style: TextStyle(
-                                        fontWeight: FontWeight.w600,
-                                        color: Colors.orange.shade800,
-                                        fontSize: 11,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                                ..._permits.map((permit) => Padding(
-                                  padding: const EdgeInsets.only(top: 2),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Expanded(
-                                        child: Text(
-                                          '  ${permit.name}',
-                                          style: const TextStyle(fontSize: 11),
-                                        ),
-                                      ),
-                                      Text(
-                                        currencyFormat.format(permit.amount),
-                                        style: const TextStyle(fontSize: 11),
-                                      ),
-                                    ],
-                                  ),
-                                )),
-                              ],
-                            ),
-                          ),
-                        ],
-
-// NEW: Show custom line items if any
-                        if (_customLineItems.isNotEmpty) ...[
-                          const SizedBox(height: 4),
-                          Container(
-                            padding: const EdgeInsets.all(6),
-                            decoration: BoxDecoration(
-                              color: Colors.purple.shade50,
-                              borderRadius: BorderRadius.circular(4),
-                            ),
-                            child: Column(
-                              children: [
-                                Row(
-                                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    Text(
-                                      'CUSTOM ITEMS:',
-                                      style: TextStyle(
-                                        fontWeight: FontWeight.w600,
-                                        color: Colors.purple.shade800,
-                                        fontSize: 11,
-                                      ),
-                                    ),
-                                    Text(
-                                      currencyFormat.format(_customLineItems.fold(0.0, (sum, item) => sum + item.amount)),
-                                      style: TextStyle(
-                                        fontWeight: FontWeight.w600,
-                                        color: Colors.purple.shade800,
-                                        fontSize: 11,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                                ..._customLineItems.map((item) => Padding(
-                                  padding: const EdgeInsets.only(top: 2),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Expanded(
-                                        child: Text(
-                                          '  ${item.name}${item.isTaxable ? '' : ' (non-taxable)'}',
-                                          style: const TextStyle(fontSize: 11),
-                                        ),
-                                      ),
-                                      Text(
-                                        currencyFormat.format(item.amount),
-                                        style: const TextStyle(fontSize: 11),
-                                      ),
-                                    ],
-                                  ),
-                                )),
-                              ],
-                            ),
-                          ),
-                        ],
-
-                        const Divider(),
-
-                        // SUBTOTAL
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            Text(
-                              'SUBTOTAL:',
-                              style: TextStyle(
-                                fontWeight: FontWeight.w600,
-                                color: Colors.blue.shade800,
-                              ),
-                            ),
-                            Text(
-                              currencyFormat.format(totalSubtotal),
-                              style: TextStyle(
-                                fontWeight: FontWeight.w600,
-                                color: Colors.blue.shade800,
-                              ),
-                            ),
-                          ],
-                        ),
-
-                        // TAX (only show if tax rate > 0)
-                        if (_taxRate > 0) ...[
-                          const SizedBox(height: 4),
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Text(
-                                'TAX (${_taxRate.toStringAsFixed(2)}%):',
-                                style: TextStyle(
-                                  color: Colors.grey.shade700,
-                                ),
-                              ),
-                              Text(
-                                currencyFormat.format(taxAmount),
-                                style: TextStyle(
-                                  color: Colors.grey.shade700,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-
-                        const SizedBox(height: 8),
-                        Container(
-                          padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                          decoration: BoxDecoration(
-                            color: Colors.blue.shade100,
-                            borderRadius: BorderRadius.circular(6),
-                          ),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Text(
-                                'TOTAL:',
-                                style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  fontSize: 16,
-                                  color: Colors.blue.shade800,
-                                ),
-                              ),
-                              Text(
-                                currencyFormat.format(totalWithTax),
-                                style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  fontSize: 18,
-                                  color: Colors.blue.shade800,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  );
-                }).toList(),
-              ),
-            ] else ...[
-              // Single-tier layout (NEW)
-              _buildSingleTierTotal(currencyFormat),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-
-// NEW: Add this helper method right after _buildQuoteTotalsSection
-  Widget _buildSingleTierTotal(NumberFormat currencyFormat) {
-    final level = _quoteLevels.first;
-    final levelSubtotal = level.subtotal;
-
-    // Calculate permits total (taxable)
-    final permitsTotal = _permits.fold(0.0, (sum, permit) => sum + permit.amount);
-
-    // Calculate custom items (separate taxable and non-taxable)
-    final taxableCustomItems = _customLineItems.where((item) => item.isTaxable).fold(0.0, (sum, item) => sum + item.amount);
-    final nonTaxableCustomItems = _customLineItems.where((item) => !item.isTaxable).fold(0.0, (sum, item) => sum + item.amount);
-
-    // Calculate subtotal before tax
-    final taxableSubtotal = levelSubtotal + permitsTotal + taxableCustomItems; // Include permits in taxable
-    final nonTaxableSubtotal = nonTaxableCustomItems;
-    final totalSubtotal = taxableSubtotal + nonTaxableSubtotal;
-
-    // Calculate tax (on taxable items including permits)
-    final taxAmount = taxableSubtotal * (_taxRate / 100);
-    final totalWithTax = totalSubtotal + taxAmount;
-
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: Colors.green.shade300),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Main product
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Expanded(
-                child: Text(
-                  '${_mainProduct!.name} (${_mainQuantity.toStringAsFixed(1)} ${_mainProduct!.unit})',
-                  style: const TextStyle(fontWeight: FontWeight.w500, fontSize: 16),
-                ),
-              ),
-              Text(
-                currencyFormat.format(level.baseProductTotal),
-                style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-              ),
-            ],
-          ),
-
-          // Added products
-          // Added products
-          ..._addedProducts.map((product) => Padding(
-            padding: const EdgeInsets.only(top: 8),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Expanded(
-                  child: Text(
-                    '${product.productName} (${product.quantity.toStringAsFixed(1)} ${product.unit})',
-                    style: const TextStyle(fontSize: 14),
-                  ),
-                ),
-                Text(
-                  currencyFormat.format(product.totalPrice),
-                  style: const TextStyle(fontSize: 14),
-                ),
-              ],
-            ),
-          )),
-
-// NEW: Show permits if any
-          if (_permits.isNotEmpty) ...[
-            const SizedBox(height: 8),
-            Container(
-              padding: const EdgeInsets.all(8),
-              decoration: BoxDecoration(
-                color: Colors.orange.shade50,
-                borderRadius: BorderRadius.circular(6),
-              ),
-              child: Column(
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        'PERMITS:',
-                        style: TextStyle(
-                          fontWeight: FontWeight.w600,
-                          color: Colors.orange.shade800,
-                          fontSize: 12,
-                        ),
-                      ),
-                      Text(
-                        currencyFormat.format(permitsTotal),
-                        style: TextStyle(
-                          fontWeight: FontWeight.w600,
-                          color: Colors.orange.shade800,
-                          fontSize: 12,
-                        ),
-                      ),
-                    ],
-                  ),
-                  ..._permits.map((permit) => Padding(
-                    padding: const EdgeInsets.only(top: 4),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Expanded(
-                          child: Text(
-                            '  ${permit.name}',
-                            style: const TextStyle(fontSize: 12),
-                          ),
-                        ),
-                        Text(
-                          currencyFormat.format(permit.amount),
-                          style: const TextStyle(fontSize: 12),
-                        ),
-                      ],
-                    ),
-                  )),
-                ],
-              ),
-            ),
-          ],
-
-// NEW: Show custom line items if any
-          if (_customLineItems.isNotEmpty) ...[
-            const SizedBox(height: 8),
-            Container(
-              padding: const EdgeInsets.all(8),
-              decoration: BoxDecoration(
-                color: Colors.purple.shade50,
-                borderRadius: BorderRadius.circular(6),
-              ),
-              child: Column(
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        'CUSTOM ITEMS:',
-                        style: TextStyle(
-                          fontWeight: FontWeight.w600,
-                          color: Colors.purple.shade800,
-                          fontSize: 12,
-                        ),
-                      ),
-                      Text(
-                        currencyFormat.format(_customLineItems.fold(0.0, (sum, item) => sum + item.amount)),
-                        style: TextStyle(
-                          fontWeight: FontWeight.w600,
-                          color: Colors.purple.shade800,
-                          fontSize: 12,
-                        ),
-                      ),
-                    ],
-                  ),
-                  ..._customLineItems.map((item) => Padding(
-                    padding: const EdgeInsets.only(top: 4),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Expanded(
-                          child: Text(
-                            '  ${item.name}${item.isTaxable ? '' : ' (non-taxable)'}',
-                            style: const TextStyle(fontSize: 12),
-                          ),
-                        ),
-                        Text(
-                          currencyFormat.format(item.amount),
-                          style: const TextStyle(fontSize: 12),
-                        ),
-                      ],
-                    ),
-                  )),
-                ],
-              ),
-            ),
-          ],
-
-          const SizedBox(height: 12),
-          const Divider(thickness: 1),
-          const SizedBox(height: 8),
-
-          // SUBTOTAL
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                'SUBTOTAL:',
-                style: TextStyle(
-                  fontWeight: FontWeight.w600,
-                  fontSize: 16,
-                  color: Colors.green.shade800,
-                ),
-              ),
-              Text(
-                currencyFormat.format(totalSubtotal),
-                style: TextStyle(
-                  fontWeight: FontWeight.w600,
-                  fontSize: 16,
-                  color: Colors.green.shade800,
-                ),
-              ),
-            ],
-          ),
-
-          // TAX (only show if tax rate > 0)
-          if (_taxRate > 0) ...[
-            const SizedBox(height: 8),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Text(
-                  'TAX (${_taxRate.toStringAsFixed(2)}%):',
-                  style: TextStyle(
-                    color: Colors.grey.shade700,
-                    fontSize: 14,
-                  ),
-                ),
-                Text(
-                  currencyFormat.format(taxAmount),
-                  style: TextStyle(
-                    color: Colors.grey.shade700,
-                    fontSize: 14,
-                  ),
-                ),
-              ],
-            ),
-          ],
-
-          const SizedBox(height: 12),
-          Container(
-            padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 16),
-            decoration: BoxDecoration(
-              color: Colors.green.shade100,
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Text(
-                  'TOTAL:',
-                  style: TextStyle(
-                    fontWeight: FontWeight.bold,
-                    fontSize: 20,
-                    color: Colors.green.shade800,
-                  ),
-                ),
-                Text(
-                  currencyFormat.format(totalWithTax),
-                  style: TextStyle(
-                    fontWeight: FontWeight.bold,
-                    fontSize: 22,
-                    color: Colors.green.shade800,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
 
 // NEW: Custom line items section
   Widget _buildCustomLineItemsSection() {
